create subvol for vms to avoid large snapshots

+++ snapper/snapper_home.sh
@@ -1,11 +1,21 @@
-echo "Moving ~/.cache to its own subvol"
-sudo btrfs sub create ~/.cache-sub
-sudo chown fblaise: ~/.cache-sub
-rsync -a ~/.cache/* ~/.cache-sub/
-mv ~/.cache ~/.cache-old
-mv ~/.cache-sub ~/.cache
-rm -rf ~/.cache-old
-sudo btrfs sub list /home
+X=$(sudo btrfs sub list /home |grep -q vms_subvol)
+if [ $X -ne 0 ]; then
+	echo "Creating subvol for virtual machines"
+	sudo btrfs sub create ~/vms_subvol
+	sudo chown fblaise: ~/vms_subvol
+fi
+
+X=$(sudo btrfs sub list /home |grep -q .cache)
+if [ $X -ne 0 ]; then
+	echo "Moving ~/.cache to its own subvol"
+	sudo btrfs sub create ~/.cache-sub
+	sudo chown fblaise: ~/.cache-sub
+	rsync -a ~/.cache/* ~/.cache-sub/
+	mv ~/.cache ~/.cache-old
+	mv ~/.cache-sub ~/.cache
+	rm -rf ~/.cache-old
+	sudo btrfs sub list /home
+fi
 
 echo
 echo "Setting up snapper for /home"
